Implement midi file playback. Add track title input.
